move wrapper metaclass to new file
